added optimization
