+++ src/Pages/main-page/main-page.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
-import React, { useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
 import { Offer } from '../../dataTypes/offer.ts';
 import { OffersList } from '../../components/offer/offers-list.tsx';
-import { Layout } from '../../components/layout.tsx';
+import { Layout } from '../../components/Layout/layout.tsx';
 import { Helmet } from 'react-helmet-async';
 import { Nullable } from 'vitest';
 import { Map } from '../../components/map/map.tsx';
@@ -14,14 +14,17 @@ import { OfferSortSelect } from '../../components/offer/offer-sort-select.tsx';
 export function MainPage(): React.JSX.Element {
   const [activeOffer, setActiveOffer] = useState<Nullable<Offer>>(null);
   const city = useAppSelector((state) => state.city);
-  const unsortedOffers = useAppSelector((state) => state.offers).filter(
-    (offer) => offer.city.name === city.name,
-  );
+  const allOffers = useAppSelector((state) => state.offers);
   const sort = useAppSelector((state) => state.sorting);
-  const offers = sort(unsortedOffers);
-  const offersCountCaption = offers.length === 0
-    ? 'No places to stay available'
-    : `${pluralizeAndCombine('place', offers.length)} to stay in ${city.name}`;
+  const unsortedOffers = useMemo(
+    () => allOffers.filter((offer) => offer.city.name === city.name),
+    [city, allOffers],
+  );
+  const offers = useMemo(() => sort(unsortedOffers), [sort, unsortedOffers]);
+  const offersCountCaption =
+    offers.length === 0
+      ? 'No places to stay available'
+      : `${pluralizeAndCombine('place', offers.length)} to stay in ${city.name}`;
   return (
     <div className="page page--gray page--main">
       <Layout>
@@ -33,14 +36,14 @@ export function MainPage(): React.JSX.Element {
             <div className="cities__places-container container">
               <section className="cities__places places">
                 <h2 className="visually-hidden">Places</h2>
-                <b className="places__found">
-                  {offersCountCaption}
-                </b>
-                <OfferSortSelect/>
+                <b className="places__found">{offersCountCaption}</b>
+                <OfferSortSelect />
                 <OffersList
                   offers={offers}
-                  onActiveOfferChange={(offer: Nullable<Offer>) =>
-                    setActiveOffer(offer)}
+                  onActiveOfferChange={useCallback(
+                    (offer: Nullable<Offer>) => setActiveOffer(offer),
+                    [],
+                  )}
                   isOnMainPage
                 />
               </section>
@@ -54,9 +57,9 @@ export function MainPage(): React.JSX.Element {
                   selectedPoint={
                     activeOffer
                       ? {
-                        location: activeOffer?.location,
-                        id: activeOffer?.id,
-                      }
+                          location: activeOffer?.location,
+                          id: activeOffer?.id,
+                        }
                       : undefined
                   }
                   isOnMainPage
