+++ src/components/offer/offer-sort-select.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { memo, useState } from 'react';
 import { Offer } from '../../dataTypes/offer.ts';
 import { SortOffers } from '../../dataTypes/sort-offers.ts';
 import { useAppDispatch } from '../../store/store.ts';
@@ -20,7 +20,7 @@ const sortingOptions: [string, SortOffers][] = [
   ],
 ];
 
-export function OfferSortSelect(): React.JSX.Element {
+function OfferSortSelectImpl(): React.JSX.Element {
   const [isOpen, setIsOpen] = useState(false);
   const [sortingOption, setSortingOption] = useState('Popular');
   const dispatch = useAppDispatch();
@@ -59,3 +59,5 @@ export function OfferSortSelect(): React.JSX.Element {
     </form>
   );
 }
+
+export const OfferSortSelect = memo(OfferSortSelectImpl);
